handled PR comments
