+++ .gitignore
@@ -40,4 +40,4 @@ yarn-error.log*
 
 # temp office files
 ~$*
-/.metadata/
+
