+++ test/cypress-e2e/cypress/e2e/2-group-application/01-organization-info-group-application-ccof.cy.js
