+++ test/cypress-e2e/cypress/fixtures/groupApplicationData.json
